Actually this couldn't work as documented, changing recommendation.
The issue is that if we point to a client-sbc, we have to open the client-sbc wide to any host because the CFA is done as a 302 (and the originator will be redirected). This might also break because of media issues.

The recommendation is thus to go through a (probably sbc-media) inbound carrier-sbc which acts as a redirect server.
(All of this means that there's little interest in using allow_onnet -- it's simpler to just have the outbound-proxy do the loopback and therefor have a single "number" record.)