inbound_enum has to be the only one for now
In order to allow inbound_enum to only bind to specific interfaces, the iptables command (in the ccnq3-inbound-enum package) would also need to know on which interfaces the re-routing must happen.

The enterprising user may decide to _not_ install ccnq3-inbound-enum and add the PREROUTING rules themselves, so that they can have both a local DNS server (such as bind9) and ccnq3-inbound-enum running on different interfaces.