With the advent of our non-TLS mediaproxy dispatcher and relay we can simplify the configuration a lot.
History: I couldn't get the 2.5.2 dispatcher to work in Debian/testing as of 2012-01-12, either using the stock (demo) certificates from mediaproxy, or my own production certificates (which work fine with an existing 2.4.4 installation).

Conclusion: Use stunnel (on the server and the client) to encrypt the connection and verify the certificates if a secure connection is required.
In ccnq3 deployments (because supporting RTP media only makes sense for registered clients) we only use mediaproxy co-located with OpenSIPS.