More code fixes.
Also FreeSwitch cannot pipe wav files. (I did my testing with PCMU indeed.)
Set the mode of the FIFO so that FreeSwitch can write to it.
Use version of pico that properly reports Etag.
Also "record" wasn't properly documented on the FS wiki, this means esp. that record_min_sec does not apply and therefor that we ought to do this ourselves. This is problematic because ideally we'd have to wait until the fifo (and CouchDB returns 201 Created) before we attempt to remove the attachment.