Made ccnq-base into a suggestion, not a requirement.
Updated the ccnq3 package so that it may still work without ccnq-base.
Especially it seems worthwhile to keep build-essential so that NPM packages may still install.
(Strictly speaking we should be using the Debian nodejs-dev package which contains those dependencies, however it is only available in sid/unstable at this time.)