opensips: do not forward REGISTER messages when unnecessary
This indicates limitations with our implementation; this is a workaround for situations where we didn't know how to deal with the REGISTER message (e.g. GRUU).
Also moved `forwarder` module at the bottom of the list of modules so that it can pick up the `do_not_forward_register` configuration flag.